spostati tutti i comandi nella classe MenuComandi
